Qwen3-VL:4B Prototype Stage 1/requirements.txt
ollama>=0.4.0
openai-whisper>=20231117
opencv-python>=4.8.0.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.14
Pillow>=10.0.0
